upload: Log all filebackend errors in UploadFromChunks
Log all the errors from a status object and not only the first one in
the exception. This assumes errors or warnings from the FileBackend
always indicate a infrastructure error and never a user-input-related
error (which should not be logged).

FileRepo::quickImportBatch merge in status of init operations and the
real operations, show all failures in the log.
Seeing only backend-fail-internal is not helpful.

Bug: T228292
Change-Id: I3f03b93de835f6f5497f4b5904b37d62d40eb9f2