Add magnify links on thumbs when not linking to file desc page
This is all in the context of !$enableLegacyMediaDOM where explicit
magnify links aren't added to the html.

If we're linking to the file description page, magnify links are already
added via css.  However, if we're not, we're missing both the url for
the media and something to click on, which can't be added with css.

Here we add the necessary url as a resource attribute on the
mw-file-element.  Parsoid already does this unconditionally so there's
precedent for it.  And, in the common case, there isn't any redundancy
since it will be omitted where mw-file-description links are present.

A script is provided to add the magnify links where needed.  These
aren't added directly to the html since they are considered purely
presentational.  Though they're styled the same as mw-file-description
links, they are given a different class, mw-file-magnify, to avoid any
confusion about their purpose and the expectation that the media element
would be found inside them.

As above, since the legacy output always includes the magnify links in
the html, it's a significant reduction in size to only be including the
resource attribute where necessary and adding the links via css.

Bug: T329413
Depends-On: Ia22eeef770980fd555fdeb83049fc0ff700be884
Change-Id: Id46d1b2ab1af3baebff13e10f1485f3cfd9a4b37