file: Check for old image existence in LocalFileDeleteBatch
Some old files are lost and have oi_archive_name = '' in the database.

In LocalFileDeleteBatch::removeNonexistentFiles a virtual url is created
with the empty string at the end. That url points to a directory and not
a file. The file backend response with a I/O error and the message
backend-fail-internal is shown,
there is no way to delete the whole file.
Skip the lost files when collecting the hash as that results in ignore

Preserve the empty archive_name on deletion/undeletion by distinct
between NULL for current file and empty string for lost old files.
Adjust api and special page a bit to not link broken file

Bug: T244567
Change-Id: I40a0e80d4be5c5c76e7685ccf86bcb3aed3e114d