RecentChanges: Remove sticky preference for "enhanced" status, make manual
By default, the toggling of the "enhanced" feature (display results
by page) was automatically saved to the user preferences. This, however
meant that the saving to preferences also happened if the user
loaded RecentChanges through a URL, which meant we changed user
preferences without the user necessarily wanting them to change.

This patch changes the behavior so that the preference is only changed
if the user explicitly and intentionally clicks the checkbox. If the
state of the checkbox changes programmatically (through visiting a
URL or any other case) then the page will update, but the user's
preference will not be changed.

Bug: T202916
Change-Id: I1c988f2a651dd42c970baf869825e4780f259b71