qunit: Remove redundant QUnit.newMwEnvironment() from testrunner.js
The "testrunner-next" test suite is an internal meta test to confirm
that `QUnit.newMwEnvironment` works correctly. The main part of this
test is in the `QUnit.module( 'testrunner' )` test suite right above
this one.

The "testrunner-next" suite exists to confirm that the env's afterEach
correctly tears down the environment to then leave clean for the next
test.

It is both redundant and actually counter-productive to define this
next test test using `newMwEnvironment`, because it means that
we're not actually testing the state after teardown. Rather, we tested
state after both the afterEach teardown *and* another beforeEach setup.

See also <https://gerrit.wikimedia.org/r/c/mediawiki/core/+/912936>

Change-Id: I54ca8674e7fb0233d2c51a823f0cb0cc9c3d54f8