Use HTMLForm and OOUI in Special:MergeHistory
Use HTMLForm with the OOUI display type in Special:MergeHistory.
This is done with minimal changes to other code, but now uses a
Status object to show errors related to the first form (the form
with the inputs for source and destination pages) inside said form.

The forms were tested under a variety of conditions and the
behaviour of the forms should not have changed.

Bug: T134675
Bug: T336531
Change-Id: I5524e0a910e6a62cabab8bc7b0fc98e1072883bd