Always Special:WantedPages vertically
Currently, when Special:WantedPages is included in another page, it renders as an inline comma-separated list rather than the vertical numbered list you get when it’s not included.  This seems inconsistent to me, since I don’t know of any other Special pages that do this, and undesirable, as I think the inline list is much harder to read in most situations.  Removing this line makes listoutput false regardless of whether it’s being included, which causes it to appear in transclusions the same way it does normally.