Don't restrict border to :not(figcaption)
Borders are only applied when figcaptions aren't visible so there's no
need to account not applying to nested media.  This helps reduce the
specificity of the default styling, making it easier to override.

To be clear, the mw-image-border class isn't applied when the format is
frame or thumb, where the caption is visible and nested media is
possible.  For the other formats, either they are a span and no
figcaption is present.  Or they are a figure, but the figcaption is set
to `display: none` in css.

Bug: T270150
Follows-Up: I70c61493fe492445702f036e5b24ef87fc3bdf43
Change-Id: I9232f5b7b9b03e2e3e9e9a6390e087bc5b0222d4