[wdio-mediawiki] Refactor waitForModuleState to async
At least in the TwoColConflict extension this method is used to
make sure that the UI is loaded and tests can be executed.

I also changed to way the arguments are passed to waitUntil, this
seem to have changed in the past. So defacto the method was called
with a different timeout default since some time and I made sure,
that that's reflected.

Bug: T337463
Change-Id: Ic1a922788d64a1c14e33bd083c60f9c7bf6eb06e