preferences search: Swap innerText → textContent for until-found compat
When a node is hidden using the `hidden="until-found"` attribute, its
`.innerText` is empty, since it corresponds (roughly) to the visible
text in the node.

Use `.textContent` instead, which just returns all text without trying
to determine if it's visible.

This affects how white space is handled (`.innerText` adds line breaks
etc. around some nodes, and `.textContent` doesn't), but our search
doesn't even try to split the text into words, so this should be okay.

Bug: T337531
Change-Id: Ie0ad950443661c4b9cc45d27119e73d340b8ec3c