rdbms: Move IDBAccessObject from /includes/dao/ to /libs/rdbms
* Is generally useful for Rdbms consumers.
* Has no dependencies besides stable constants and interface from Rdbms.
* Is generic, i.e. not specific to MediaWiki in its naming, purpose,
  or required domain knowledge.

Current rationale is that "dao" is not a recognised component
in MediaWiki core. Its super tiny and a handful of changes in over a
decade, so seems overkill to make a big deal out of it by creating a
whole new component for it in all the usual places. Yet, having it
fall into a gap with no owner for issue tracking, code review, no
documentation, and no definition of what this directory is for
exactly, is also counter-productive. Hence, finding a suitable place
for it.

Change-Id: I8cbb9ebfa8873085a30cb355e3299493538006ae