HTMLForm: Allow conditional prefilling even without identifier
Special:Logs is prefilling fields manually by the `default` param
for compatibility with links without `wpFormIdentifier` set.
That disabled the validation built in HTMLForm, since we would
give up when the default value is invalid.

Prefilling fields and trying submit should be fine if the caller
is sure that the page only contains that GET form.

Bug: T338042
Change-Id: I047d2546d3dacc672304dab1e362c08992e68308