api: Silence database write warnings from ApiLogout
At WMF, we get about 36,000 pre-send warnings from TransactionProfiler
(Logstash: mediawiki-rdbms-performance), of which 9000 are from
ApiLogout.

For example:

> Expectation (writes <= 0) by ApiMain::setRequestExpectations not met
> (actual: 1):
> role-primary: INSERT INTO `cu_changes` (…) VALUES (N,...,N,'X',N,...,N)

These are actually fine as-is given that api?action=logout already
requires POST method and CSRF tokens. We just didn't indicate yet
that it was expected to perform writes (as opposed to using POST for
cache reasons or due to query size).

Change-Id: I89aa7bf672fbf71f153971ad9c0a5d6fbb7733f5