SpecialMovePage: Avoid TransactionProfiler warnings
SpecialMovePage currently implicitly uses RIGOR_SECURE for some
permission checks when rendering the initial form, which causes
TransactionProfiler warnings to be logged. So, switch the main move
permission check to use RIGOR_FULL instead so that it still checks user
blocks but from a replica, rather than a primary database. Also use
quickUserCan() to determine whether to render the "move subpages"
checkbox, for consistency with the rest of the rendering logic.

Note that this can still result in TransactionProfiler warnings if an
blocked user attempts to access the special page and gets autoblocked as
a result.

Bug: T283265
Change-Id: I6cf7da775c7727aca56de904af97e356df556f65