jquery.makeCollapsible: Use `unset: all` on buttons
Instead of trying to unset individual properties, just unset everything:
the appearance of the button should resemble the appearance of a <span>,
and <span>s appear exactly how the defaults appear. Most notably, the
previous code unsetting individual properties didn’t unset the
foreground color, which caused it to use the browser default color
instead of the inherited color.

Also remove `-ms-user-select`, because it’s used only by Internet
Explorer and Edge Legacy, neither of which get JavaScript anymore
(this stylesheet is loaded only when JS is available).

Bug: T333357
Change-Id: Iac1fcb9a9beb0536e3848ae98b110cc6192bb178