tests: Check that extension service getter methods exist
We had some tests for the structure of all existing methods, but nothing
actually asserted the existence of the methods based on the service
names. See the Depends-On for some services that were still missing
their methods.

Implemented as a single test, not using a data provider, because the
service container isn’t available in data providers (nor should it be,
see T332865). The $serviceNamesWithoutMethods mechanism is needed for
AbuseFilter, where AbuseFilterRunnerFactory is a backwards compatibility
alias for AbuseFilterFilterRunnerFactory.

Change-Id: If5af88e7f70b83d53f66b9617a5ef37daf81830f
Depends-On: I22c1b60eb014c3ca75079c99a44592321e71d9e9
Depends-On: I28ad8cf8003cac07f5ad1f7bbd7d7f52e34a4ed0
Depends-On: Idedb87e64a6df02b0edae8d9e7dbf441752dc480