Set default accesskey+tooltip for subject namespace tabs
Until now, accesskeys and tooltips for the subject namespace have
had to be set manually by creating or defining
[[MediaWiki:Accesskey-ca-nstab-<ns>]] and
[[MediaWiki:Tooltip-ca-nstab-<ns>]] for each namespace; this also
applies to all extra (non-standard) namespaces, as well as extra
namespaces defined by extensions.

This change sets a default accesskey and tooltip for the subject
namespace tab if no accesskey has been defined for it. The accesskey
is set to "c" by default, because that is the accesskey that was used
for all default namespaces, except – inexplicably – the project
namespace, which had "a". This change also changes the project
namespace's accesskey to "c" to be in line with the rest of the
namespaces.

Bug: T22126
Change-Id: I3269d3b4fe18a4f3fcf108435cf1bbaffbeac1b1