HookContainer: allow registering broken callables.
Extensions may implement hooks defined by other extensions.
When the other extension is not present, the base class or interface of
the hook handler may also not be present. In that case, we should still
allow the handler to be registered. This is harmless, since the hook
will never fire if the extension that defines it is not loaded.

Bug: T339834
Change-Id: Id975370dececcb09ae5d00e6f68da57dc22fb267