Do not generate diffs for slots with identical content
When the old and new revisions have identical content in a given slot,
do not include the relevant SlotDiffRenderer in the return value of
DifferenceEngine::getSlotDiffRenderers(), so that the relevant table
prefixes and modules will be omitted.

Also:

* In TextSlotDiffRenderer::getTextDiff(), return an empty string when
  the inputs are equal, instead of invoking the diff engine. Previously,
  the inline format would produce a <tr> element in this case, causing
  the diff-empty message to be omitted.
* Clarify the doc comment up the stack, indicating SlotDiffRenderer may
  return zero <tr> tags, it doesn't have to be one or more.
* Fix the documented return type of getSlotContents() so that @var is
  not needed in the caller.
* Don't convert the return value of SlotDiffRenderer::getDiff() to
  boolean. It's always a string so compare it to the empty string.

Note that the cache key for inline empty diffs will change, because
TextSlotDiffRenderer::getExtraCacheKeys() is not called anymore for
empty diffs.

Bug: T338670
Change-Id: I0774ba0b159ac43ec214403cf2d06740f6d067cd