Fix image size in the last row of packed galleries
In packed galleries, the size of images can be increased up to 1.5x to
fill a full row. The last row, however, is often too short to do that.

In change 3c87c311cf36689de8562b708bde36697e3f726e (2013), it was
intended for the last row to use the average scaling factor of all
other rows, but due to a simple mistake the row was instead always
using the default size of 1x.

Testing on an example gallery at
<https://www.mediawiki.org/wiki/Phabricator/Help#File_visibility>:
* Before: https://phabricator.wikimedia.org/F37118396
* After: https://phabricator.wikimedia.org/F37118397

Change-Id: Ia2e4536c6b15200a071fa983468c3856785b14b4