Add SelectQueryBuilder::merge() and ::clearFields()
Add SelectQueryBuilder::merge(), which is essentially the same as

  $sqb1->queryInfo( $sqb2->getQueryInfo() );

This allows things like User::getQueryInfo() to be replaced with similar
methods that return a pre-configured SelectQueryBuilder for merging into
some larger query.

Add SelectQueryBuilder::clearFields(), which erases the list of fields,
allowing queries to be merged without merging their fields, which is a
common use case.

Change-Id: If1b36acbc6aebec472fa83dd9481295f0648bc61