DoubleRedirectJob: Inject appropriate service into job p.1
As a first pass, inject various services into the job as this is
now supported by JobFactory.

The second pass will have to survey callers of fixRedirects() and
pass in various services that would be used in the static method.
Also, we need to do same for getFinalDestination() which needs a
DBLoadBalancerFactory service.

Bug: T265942
Change-Id: If61a32b6ce32dc934533603a0e99414ab5771909