preview: Insert template list after it's all built, rather than before
The live-preview template list was being inserted while it was being
queried, leading to the length of the list changing dynamically and
the page visually moving around. To avoid this, this change holds
the new template list until it's all been built, and then inserts it
at that point.

Also fixes an issue with the collapsibility of the template list.

Bug: T337280
Change-Id: I828d1a72f751b221c9da56e13d846f0ee2f2e281