languages: Prettify ZhConversion.php
Update the python script so that the generated PHP file uses proper
indentation. Correct indentation will be enforced with PHPCS (T319352),
and this file alone is responsible for roughly 20000 violations. Even if
it's automatically generated, to me it seems better to have it comply
with our conventions, rather than excluding it.

Bug: T319352
Change-Id: Iafaca207d21e56257d2e0adec53934855a5a7c2a