specials: Simplify query on Special:ListRedirects
The column redirid of the result is not used in the code,
remove the column and the obsolete join for the value.
The old query is also wrong as it does not use rd_interwiki = '' or NULL
to avoid getting a page id for interwiki redirects.

The redirid exists since the querypage-work2 branch was merged into
master (56b5aa18)

Change-Id: I0f05ec4a75cc298034a9d79b8200ba610c8a3567