phpunit: Deprecate tests/phpunit/phpunit.php script
`composer phpunit:entrypoint` should be used instead. Copy all the
setup to a new bootstrap file, with some adjustments for the bootstrap
not being in the global scope when PHPUnit loads it.

Leave the old script and bootstrap in place (e.g., for CI), but make
them print deprecation warnings.

Update the composer scripts to use the standard `phpunit` entrypoint.

The bootstrap code was copied almost verbatim from the phpunit.php
entrypoint to minimize changes. Any improvements can be done in other
patches.

Bug: T90875
Change-Id: I95a80c44f8d88c3c498efe1ae64008f0ff1eea55