Fix requireOnceInGlobalScope for PHP < 8.1
In PHP < 8.1, the $GLOBALS variable doesn't follow the usual array
by-value semantics: $originalGlobals was actually a reference to the
$GLOBALS array. This means that its value might be different before and
after the require_once. See https://3v4l.org/BeUgm.

In turn, this leads to a bug when trying to replace phpunit.php, where
Wikibase's $wgWBClientSettings is set to the empty array by an extension
function. In turn, this causes the variable to be ignored by this method
because it wasn't in the original list of globals (and therefore it
wasn't put into the global state), but it was in $originalGlobals after
the require_once, which excluded it from being exported later. With that
variable's value missing, some tests that needed specific config were
being skipped.

Again, note that this is only an issue in Wikimedia CI when running PHP
7.4 jobs, and not elsewhere if PHP >= 8.1 is used (for instance, in
MediaWiki-Docker).

Bug: T90875
Change-Id: Ib1954ecf5a01df9b55c4d55cbeb61c7737d51f9f