Do not parse trivial HTML as wikitext on Special:Version
This showed up as a major time sink in a profile run I did. What this
code does is turning the two fields "homepage" and "name" into a
wikitext snippet and do a separate wikitext parser run just to create
an <a> element. This adds up because it happens almost 300 times.

Instead we can create the <a> element directly.

External libraries are not expected to contain wikitext in a "name"
field.

Change-Id: Iba6ecf68d527f11d81be3f2dcc2d968a75a9db70