Don't parse HTML as wikitext on grants/rights special pages
These pieces are known to be trivial HTML. The content is known to
be a plain text identifier. We don't need to parse it again as if it
is wikitext. rawParams() is the way to mark such parameters as "show
as is, don't parse again".

This also removes some double escaping that was added as part of
Ida2b686 in 2015. It was most certainly never doing anything because
the incoming identifiers are plain ASCII with nothing to escape.

Change-Id: I74fcdab3bbd9fb4f692176a6ac4cef6202db7f7e