Minimize code duplication in StatusValue methods
Use modern array destructuring and such to make the code – so I
hope – much more readable.

Also fix a mistake in the Status test setup. This caused the mocked
Message::getParams() to return an array with string keys. This is
impossible in non-mocked scenarios. There is an array_values() call
in the Message constructor that doesn't allow string keys.

Change-Id: Ief09feb06c60d3ffbfa95fa8ef56b4a9829da680