Use special pages instead of actions for revision delete and edit tags
Timo mentioned the difficulty to port these special pages to actions
while with use cases of log events and oldimages, and pointed out we
may re-consider the direction of the task.
They indeed can be switched to use special pages directly without
any issue, so let's go for this direction.

Bug: T323338
Bug: T340853
Change-Id: I1947c1d4ccfb624640de761f120b3f1a482de740