ExternalLinks: Make order by and continue only rely on el_id in READ NEW
This way of doing continue was built when when externallinks didn't have
el_to as an auto_increment PK. Now that it has had something to use for
continue, it doesn't need to rely on those fields in order by nor
continue. These fields are set in the query and order by on a field
that's already specified in WHERE condition is not useful nor gives any
value.

Bug: T341000
Bug: T47237
Change-Id: I5c3bea57792bdce0ea611c3b33b4259b8cafc9d0