mediawiki.user: Validate the cookie in sessionId() before returning
Other code depends on the return value being exactly 20 hex
characters, but the cookie value is user-controlled and could be
anything, and we also used different formats in the past ourselves,
which can still persist in cookies.

As a bonus, cache the value in a local variable after validating
or generating it, instead of reading the cookie every time.

Bug: T283881
Change-Id: I0ea1c87e46e9a1de571f3895cff08d3e802b0362