Ignore parameters to OutputPage::setupOOUI() and deduplicate calls
There are about 33 callers in Code Search, and almost all were calling
it with incorrect parameters.

It was incorrect to call it with no parameters, since this would
globally set the wrong skin and direction. Most callers were doing this.

It was incorrect to call it with the values from some local context.
We are setting a global singleton here so only the global RequestContext
has the correct values.

Usually, setupOOUI() is called to avoid the exception which is thrown
when HTML is generated without a singleton. But it was often called
multiple times in a request, because different parts of the stack need
OOUI HTML and they don't know whether OOUI has been initialised.

So, ignore the parameters passed to setupOOUI(). Ignore all calls after
the first one, so that we can stop throwing away the theme object. Reset
the cached data when RequestContext::setSkin or
RequestContext::setLanguage() are called.

Add RequestContext::getSkinName() so that the name of the skin can be
determined without actually creating the skin object. It would be
complicated to add the method to IContextSource since there is a
method of the same name in Skin, which implements IContextSource.

Change-Id: I7a1e8b613408d2ecc2e497e0672644af85ce0da3