RecentChanges: Add process caching to RecentChanges UserLinks and Tags parsing
Rendering UserLinks or Tags requires parsing, which when run on thousands of records, adds up to a significant amount of processing time. More often than not, this can be optimized by storing already visited tags or user links as those "renders" will not change for the single request run and are repeated constantly on the list.

Bug: T341319
Change-Id: I91588aebae7e49e3d3cb77702cf28677b4a14c8d