Fix TestLocalisationCache being way to small
I dug into a particular profile via KCachegrind (specifically the
profile for DirectParsoidClientTest in VisualEditor) and found this
surprising bottleneck. Let's say a test uses a data provider to test
a particular piece of code with 5 different languages. And then
another piece of code with 5 languages. And so on. Because the cache
can only hold 4 languages there will never be a successfull cache
hit.

The items in the cache are something like 1 or 2 megabytes each. It
should be fine to hold 16 of them in memory. Possibly even more. The
sweet spot in my example was 10. This is when the runtime of
DirectParsoidClientTest gets suddenly cut in half.

Bug: T225730
Change-Id: Ie04f338ee8a59f08be50947d794626ed52d793bc