Use explicit title in RevertedTagUpdateIntegrationTest
Previously, the test implicitly relied on getExistingTestPage() always
returning the same page. IMHO this makes the test harder to understand
even when the assumption is true, and all the setupEditsOnPage() callers
already have a $pageTitle available in the immediate vicinity of the
call, so let’s just pass $pageTitle into setupEditsOnPage().

Bug: T341344
Change-Id: Icd1a97329466189722a8f6051dace045ae1edd92