Add a config to stop shipping the legacy media styles
Since the default for wgParserEnableLegacyMediaDOM is now `false`, this
config will also be set to `false` and wikis can enable it while it's
still necessary.

The plan will be to gradually roll this change out, as was done for
wgParserEnableLegacyMediaDOM, to develop a migration path for templates
and extensions that mimic the legacy media output.

We can start on officewiki.

Bug: T318433
Change-Id: I70e537b2beb99de9a41e7dc66a0d60a508009887