jobqueue: Use a proper config system for dealing with settings (p.1)
WANObjectCache has already been injected, just use it instead via the
member variable. Also, make use of a proper config system like
GlobalVarConfig for accessing global variables and MainConfigNames
to pass in JobClasses setting.

Centralize checking for job beings added to the queue if they can't
be run because they're not on the correct wiki. Let the factory
deal with this and avoid the JQG to know about it. Also, centralize
logic for getting cached job classes in the factory.

NOTE: private constant CACHE_VERSION is not used within the class,
so removed it. It's private so easy to remove directly.

Change-Id: Ia61e00d6dc98c206509724413cf641a225649dbd