tests: Avoid DB queries in ApiStructureTest::provideParameters
Some of the methods called in the data provider, like getAllowedParams,
may make read queries. This should be avoided in data providers (see
e.g. T312849). Instantiate the module in the test method instead, and
process all parameters there. This does mean that a failure with a
single parameter will make the test end without testing the remaining
parameters, but there isn't much we can do about it.

Bug: T341731
Change-Id: I7eeed0da5da61e90c6fbfa0411a2b3b416395167